requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
certifi==2022.9.14
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-rest-auth @ file:///home/conda/feedstock_root/build_artifacts/django-rest-auth_1637178492924/work
djangorestframework @ file:///home/conda/feedstock_root/build_artifacts/djangorestframework_1663854154474/work
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pytz @ file:///C:/Windows/TEMP/abs_90eacd4e-8eff-491e-b26e-f707eba2cbe1ujvbhqz1/croots/recipe/pytz_1654762631027/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
wincertstore==0.2
